+++ js/utils/get-random-positive-integer.js
@@ -28,3 +28,4 @@ function getRandomPositiveInteger (a, b) {
   // потому что Math.random() генерирует только дробные числа и ноль.
   return Math.floor(result);
 }
+export default getRandomPositiveInteger;
